[clang-tidy] add deduplication support for run-clang-tidy.py
Summary:
`run-clang-tidy.py` is the parallel executor for `clang-tidy`. Due to the
common header-inclusion problem in C++/C diagnostics that are usually emitted
in class declarations are emitted every time their corresponding header is
included.

This results in a *VERY* high amount of spam and renders the output basically
useles for bigger projects.
With this patch `run-clang-tidy.py` gets another option that enables
deduplication of all emitted diagnostics. This is achieved with parsing the
diagnostic output from each `clang-tidy` invocation, identifying warnings and
error and parsing until the next occurence of an error or warning. The collected
diagnostic is hashed and stored in a set. Every new diagnostic will only be
emitted if its hash is not in the set already.

Reviewers: alexfh, aaron.ballman, hokein, sammccall, lebedev.ri

Subscribers: mgorny, xazax.hun, cfe-commits

Differential Revision: https://reviews.llvm.org/D54141