[clang-tidy] refactor bugprone-exception-escape analysis into class
Summary:
The check `bugprone-exception-escape` does an AST-based analysis to determine
if a function might throw an exception and warns based on that information.
The analysis part is refactored into a standalone class similiar to
`ExprMutAnalyzer` that is generally useful.
I intent to use that class in a new check to automatically introduce `noexcept`
if possible.

`bugprone-exception-escape` uses the consistent semicolon-separated list for
configuration now as well.

Reviewers: aaron.ballman, alexfh, hokein, baloghadamsoftware

Subscribers: mgorny, xazax.hun, rnkovacs, cfe-commits

Differential Revision: https://reviews.llvm.org/D57100